requirements.txt
torchvision
opencv-python
numpy
xformers
Pillow
DateTime
matplotlib
plyfile
HTML4Vision
timm
tensorboardX
imgaug
iopath
imagecorruptions
mmcv
pytorch-tcn
mmengine